polish: split dev function out of index.tsx
